Move Traceable trait to mozjs-sys.
