+++ mozjs-sys/src/lib.rs
@@ -20,12 +20,14 @@ pub mod glue;
 pub mod jsgc;
 pub mod jsid;
 pub mod jsval;
+pub mod trace;
 
 // Reexport the bindings in the jsapi module
 pub use crate::generated::root as jsapi;
 
 // The bindings generated by bindgen
 #[doc(hidden)]
+#[allow(dead_code)]
 mod generated {
     include!(concat!(env!("OUT_DIR"), "/build/jsapi.rs"));
 }
